+++ src/_types/index.d.ts
@@ -1,3 +1,14 @@
+import type { error_o_T } from '@ctx-core/error'
+export interface gtm_body_script_html_opts_T extends error_o_T {
+	GTM_ID?:string
+	GOOGLE_TRACKING_ID?:string
+	GA_ID?:string
+}
+export interface gtm_script_html_opts_T extends error_o_T {
+	GTM_ID?:string
+	dataLayer?:[]
+}
+export declare type gtm_script_html_opts__T = gtm_script_html_opts_T
 export interface WebFontConfig {
 	loading?:()=>void,
 	active?:()=>void,