+++ src/dataLayer__init/index.d.ts
@@ -1,6 +1,6 @@
-import type { gtm_script_html_opts__T } from '../gtm_script_html_opts___T'
-export declare function dataLayer__init(opts?:gtm_script_html_opts__T):void;
+import type { gtm_script_html_opts_T } from '../_types'
+export declare function dataLayer__init(opts?:gtm_script_html_opts_T):void
 export {
 	dataLayer__init as init_dataLayer,
-	dataLayer__init as init__dataLayer
+	dataLayer__init as init__dataLayer,
 }