+++ src/gtag_html_/index.d.ts
@@ -2,8 +2,9 @@ import { error_o_T } from '@ctx-core/error'
 /**
  * Html to add gtag.js to the site
  */
-export declare function gtag_html_(opts?:_gtag_html_opts_T):string
+export declare function gtag_html_(opts?:gtag_html_opts_T):string
 export { gtag_html_ as _gtag_html, gtag_html_ as _html__gtag, }
-export interface _gtag_html_opts_T extends error_o_T {
-	GOOGLE_TRACKING_ID?:string;
+export interface gtag_html_opts_T extends error_o_T {
+	GOOGLE_TRACKING_ID?:string
 }
+export declare type _gtag_html_opts_T = gtag_html_opts_T