+++ src/gtm_body_script_html_opts_T/index.d.ts
@@ -1,6 +1,0 @@
-import type { error_o_T } from '@ctx-core/error'
-export interface gtm_body_script_html_opts_T extends error_o_T {
-	GTM_ID?:string
-	GOOGLE_TRACKING_ID?:string
-	GA_ID?:string
-}