+++ src/gtm_head_script_html_/index.d.ts
@@ -1,6 +1,9 @@
-import type { gtm_script_html_opts__T } from '../gtm_script_html_opts___T'
+import type { gtm_script_html_opts_T } from '../_types'
 /**
  * Google Tag Manager script html to place at the top of `<head>`
  */
-export declare function gtm_head_script_html_(opts?:gtm_script_html_opts__T):string;
-export { gtm_head_script_html_ as _gtm_head_script_html, gtm_head_script_html_ as _html__script__head__gtm, }
+export declare function gtm_head_script_html_(opts?:gtm_script_html_opts_T):string
+export {
+	gtm_head_script_html_ as _gtm_head_script_html,
+	gtm_head_script_html_ as _html__script__head__gtm,
+}