+++ src/gtm_script_html_/index.d.ts
@@ -1,11 +1,8 @@
-import type { gtm_script_html_opts__T } from '../gtm_script_html_opts___T'
+import type { gtm_script_html_opts_T } from '../_types'
 /**
  * Html to add gtm.js to the page
- * @param opts
- * @param opts.GTM_ID
- * @returns {string}
  */
-export declare function gtm_script_html_(opts?:gtm_script_html_opts__T):string;
+export declare function gtm_script_html_(opts?:gtm_script_html_opts_T):string
 export {
 	gtm_script_html_ as _gtm_script_html,
 	gtm_script_html_ as _html__script__gtm,