+++ src/gtm_script_html_opts___T/index.d.ts
@@ -1,5 +1,0 @@
-import type { error_o_T } from '@ctx-core/error'
-export interface gtm_script_html_opts__T extends error_o_T {
-	GTM_ID?:string;
-	dataLayer?:[];
-}